requirements.txt
robotframework>=7.0.0
robotframework-browser

python-dotenv